fix(autobot_shared): move retry_mechanism into autobot_shared to fix slm-agent startup crash (MVA-1493)
`connection_manager.py` hard-imported `retry_mechanism` from the
top-level, which lives only in `autobot-backend/`. Any process that
imports `autobot_shared.redis_management` without `autobot-backend` on
its path (e.g. the SLM-manager backend) crashed at startup with
`ModuleNotFoundError: No module named 'retry_mechanism'`.

Fix:
- Add `autobot_shared/retry_mechanism.py` — canonical home of
  `RetryConfig`, `RetryMechanism`, `RetryStrategy`, and related helpers.
  Only change vs. the original: imports now come from
  `autobot_shared.ssot_constants` instead of `constants.threshold_constants`.
- Replace `autobot-backend/retry_mechanism.py` with a thin re-export
  shim so all existing callers in autobot-backend continue to work
  without modification.
- Update `connection_manager.py` imports to use
  `autobot_shared.retry_mechanism` and move them into the proper import
  block.
